added ownership to RPC server

+++ src/stateless/trex_stream_api.h
@@ -90,15 +90,15 @@ private:
  */
 class TrexStreamContinuous : public TrexStream {
 public:
-    TrexStreamContinuous(uint8_t port_id, uint32_t stream_id, uint32_t pps) : TrexStream(port_id, stream_id), m_pps(pps) {
+    TrexStreamContinuous(uint8_t port_id, uint32_t stream_id, double pps) : TrexStream(port_id, stream_id), m_pps(pps) {
     }
 
-    uint32_t get_pps() {
+    double get_pps() {
         return m_pps;
     }
 
 protected:
-    uint32_t m_pps;
+    double m_pps;
 };
 
 /**
@@ -107,7 +107,7 @@ protected:
  */
 class TrexStreamBurst : public TrexStream {
 public:
-    TrexStreamBurst(uint8_t port_id, uint32_t stream_id, uint32_t total_pkts, uint32_t pps) : 
+    TrexStreamBurst(uint8_t port_id, uint32_t stream_id, uint32_t total_pkts, double pps) : 
         TrexStream(port_id, stream_id),
         m_total_pkts(total_pkts),
         m_pps(pps) {
@@ -115,7 +115,7 @@ public:
 
 protected:
     uint32_t   m_total_pkts;
-    uint32_t   m_pps;
+    double     m_pps;
 };
 
 /**
@@ -127,7 +127,7 @@ public:
     TrexStreamMultiBurst(uint8_t  port_id,
                          uint32_t stream_id,
                          uint32_t pkts_per_burst,
-                         uint32_t pps,
+                         double   pps,
                          uint32_t num_bursts,
                          double   ibg_usec) : TrexStreamBurst(port_id, stream_id, pkts_per_burst, pps), m_num_bursts(num_bursts), m_ibg_usec(ibg_usec) {
